Add files via upload
This is my project submission of DATA COOMUNICATION AND NETWORK LAB. This was a project done in a group of two. 
We have designed a multi-branch network topology intended for a hypothetical organization with branches in multiple cities. The evaluation of this design focuses on its robustness, scalability, and efficiency in handling inter-branch communications and services.